Add Swagger

+++ src/main/java/com/api/patientApi/repositories/UserRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserModel, Long> {
+public interface UserRepository extends JpaRepository<UserModel, Integer> {
 
     Optional<UserModel> findByUsername(String username);
 }
